Clarify that pcap_lookupnet() does *NOT* return an IP address.
Note that what it returns is the network number, i.e. the IP address
ANDed with the network mask, *NOT* the IP address.  A number of people
have been confused by this.  Also, use "network mask" rather than "net
mask", and note that pcap_lookupnet() returns an IPv4 network number and
mask and that it only returns one of them (the interface might have
multiple IPv4 addresses).