Reflect name changes to some LINKTYPE_ values.
To match the changes made in the libpcap source:

Rename LINKTYPE_TOKEN_RING to LINKTYPE_IEEE802_5 to clarify that it
applies to IEEE 802.5 token ring, not to other token rings.

Give all the 802.11+radio metadata LINKTYPE_ values names that begin
with LINKTYPE_IEEE802_11, and rename LINKTYPE_IEEE802_11_RADIO to
LINKTYPE_IEEE802_11_RADIOTAP to clarify that it's radiotap.