regenerate man pages
This round has been done on Ubuntu 16 where the man tool does not
produce the extra newlines in the .txt version anymore, such that the
contents of the files is now closer to the man pages as they used to
appear in interactive shell sessions. Also man2html output includes a
trailing whitespace in one case.