Use consistent style for the /release autoindex.
Follow up from commit fd1bf66 and use the stylesheet for /release/ too.
Add IndexOptions without +SuppressSize and +IgnoreClient because for this
particular directory file size is a meaningful column. Using or not using
-FancyIndexing +HTMLTable does not seem to make any difference, but leave
these in place for consistency with the other .htaccess files. Amend the
stylesheet to pad the columns a bit for readability.