Declutter the page menus a little bit.
Links to mirrors.html and faq.html were (almost) only in index.html,
move these to the top menu.

Links to index.html#source, index.html#latest-releases,
index.html#mailing-lists and index.html#contribute were both in the
sidebar menu and in the top menu, leave them in the sidebar menu only.
Make an exemption for index.html#security, which definitely needs to be
in the top menu (that is, on every page), possibly as a standalone page
later.