Make "old releases" a standalone page.
This information is not useful to most visitors, but it currently
occupies quite a lot of screen space on the index page. Make it a
top-level page, where it should be possible to provide as much detail
for as many previous releases as necessary (maybe potentially even add
past milestones to make it a project timeline).