Add some LINKTYPE_* symlinks for clarity.
GH pull request #14 states the problem that in linktypes/ a page name
is usually based on the link type it specifies. Two pages used not to
follow this convention because they specify two link types each. Solve
this problem by adding symlinks and telling mod_autoindex not to list
the original files, so from the web client point of view there is a
standalone page for each link type.