Lose Dublin Core [no-op] meta information.
Dublin Core is one of the things originally invented by good people for
good people to tell what a document is, and later abused by bad people
to lie that a document is something that it actually is not. Internet
search engines have been using their own means to comprehend the web for
many years. I know this because I had worked in a search engine company
for 7 years, but if anybody wants to verify the claim in a shorter time,
here is a research article on the subject, published 15 years ago:
http://www.webology.org/2005/v2n2/a13.html