regen_manpages.sh: Spell "man page", not "manpage".
There's no such word as "manpage", so when in commit 6907b38 I changed
the output in Ubuntu to say "manpage" instead of "man page" to be
consistent with the output in Fedora, that was exactly the opposite of
the right thing to do. Try to get it right on this occasion.