Lose other A elements inside H1 and H2.
In faq.html and index.html the anchors have a purpose, convert them to
"id" attributes so they don't get in the way of switching to today's
snapshot of the HTML spec.  In linktypes.html, related.html and
season-of-docs.html the anchors have no purpose, delete them.  Update
style.css to match the simplified headers.