Put markup right in pcap.html.
Use the same arrangement of DIV elements in place as in linktypes.html
to have the styles apply as intended.  Add lots of CODE elements.  Add
hyperlinks to some libpcap man pages.  Use entities within PRE code
blocks.  In the table use TH for the headers and replace all attributes
with a class that more or less serves the purpose.  Lose some trailing
space.