Claim pages to be today's snapshot of HTML spec.
The UTF-8 charset declaration I made in commit 7a2491e is invalid for
HTML 4.01.  Fixing that would require either switching back to
http-equiv META or using a spec where it is valid.  Do the latter, as it
clears some other backlog.  Similar work remains to be done for the
generated man pages, which still use HTML 4.01.