Refine assorted hyperlinks.
In index.html refer to README and installation notes in the
"documentation" section, use a local copy of ethereal-tcpdump.pdf, add
two new external resources to the list.  In the "license" section add a
link to 2-clause BSD license for completeness.  In the "current
development versions" condense the hyperlinks to project respositories.

Link to public-cve-list.txt from security.html, not from the "latest
releases" section of index.html.

In the footer update the URL of Luis MartinGarcia web site.