Move the CI digest to a page of its own.
The long list of CI badges in the middle of the index page takes too
much vertical space and makes other, otherwise interrelated sections
difficult to follow.  Since the CI digest is not of interest to most web
site visitors, move it to a new page (ci.html) and cross-link the new
page and the index page.