Make HTML and CSS files expire. [skip ci]
When an .html file has changed on the server, my Firefox manages to
detect it most of the time when it asks the server about the document
and sees a newer Last-Modified header value.

However, it does not seem to ask the server about the associated .png
and .css files, so continues to use cached versions of these with the
updated .html file.  Whilst the former do not change, the latter do, and
the pages look wrong until a forced reload.

Add initial directives for mod_expires to see if it helps.