Remove BPF for Ultrix 4.2.
The README in bpfext42.tar.Z says:

"Unofficial Field Test Release of
Berkeley Packet Filter Extensions for Ultrix

Version: 22 July 1991"

The last version of Ultrix (4.5) was released in 1995.  As far as README
and INSTALL files go, libpcap does not claim Ultrix support since recent
commit 1fc034d, and tcpdump has not claimed Ultrix support for a while.
It is time to remove this file.