Restyle HTML man pages on the fly. [skip ci]
Add a PHP script to amend HTML man pages so they look more in line with
the rest of the web site and display other available versions in the
version boilerplate.  Add mod_rewrite directives to have the script
handle the existing URLs.  This way the existing workflow remains in
place and it is possible to view the generated files from a filesystem,
as before.