Lose the [Valid X] badges in some pages.
The [Valid CSS] badge is excess, as all CSS in this web site is in two
files and can be validated without having the badge in every page's
footer.

The [Valid HTML] badge currently works only for index.html and the man
pages, as the current version of validator.w3.org fails to process the
referrer URL if the page uses a snapshot of the HTML spec (e.g. if you
clicks on the badge at https://www.tcpdump.org/linktypes.html, you get
validation results for https://www.tcpdump.org/).

Remove these badges from the files generated with regen_html_pages.sh.