Fix restyle.php after commit c5a393a. [skip ci]
The preg_replace() substitution did not remove the directory prefix if
it was present.  This did not work right for libpcap releases, for
example, for /manpages/libpcap-1.9.0/pcap_datalink.3pcap.html the
generated alternate href was "libpcap-1.9.0/pcap_datalink.3pcap.txt",
which resolves to a non-existent path.

Just derive the href from the $basename and simplify the code.