Address assorted issues in HTML pages.
Remove the reference to pcap_datalink_ex(), which has been in linktypes.html
since the initial git revision.  The function does not exist, and it is not a
misspelling of pcap_datalink_ext(), which returns the data that is clearly *not*
the assigned DLT_ or LINKTYPE_ value.

Lose duplicate text in FAQ Q8. Fix some markup, spelling and punctuation.