Neutralize 36 invalid links in HTML man pages. [skip ci]
As it turns out, older man pages refer to ethers(3N) and usermod(1M), as
well as to pcap-tstamp-type(7), which does not exist and was replaced
much later.  Also man2html overshoots trying to find man page references
in old tcpdump output example.  Convert these occurrences back to plain
text so they do not generate HTTP 404 in the logs.