Make /release/signing-key.asc a redirect. [skip ci]
This is a follow-up to commit 106dca3.  When the current revision of the
web site refers to a PGP public key, most of the time it refers by key
ID (the only exception is autoindex contents).  Other links to the old
document, which has been around for a long time, are elsewhere on the
Internet, and these should work for quite a while longer.

However, instead of pretending there is a standalone document (via the
symlink) let the clients see that the old URL is no longer the most
appropriate, and remove the old filename from autoindex to stop telling
the crawlers the old document exists.