restyle.php: Generate better page titles.
Start the HTML document title with the man page name, this way as much
of the latter as possible appears in the browser tab title instead of
"Man page of... [x]".  Use lower case for the page name and add the
section in upper case, this way the title style is consistent with the
document body.