BPF Exam: Make assorted minor improvements.
Add a comment to explain Radare2 installation on the web-server.  In the
overview reword some prose for clarity and add more hyperlinks.

In pipe_process() rename the 2nd argument and make it optional.  In
run_filtertest() add type declarations.  Escape the output of
run_caper().  If there is an exception, print the error message before
the normal output, not after -- this way it is easier to see.