ethereal-tcpdump.pdf is badly out date, remove it.
The contents is a two-column represenation of pcap-filter(7), which has
been available online for quite a while, with many additions and
corrections.

The reference first appeared on www.tcpdump.org shortly before
2007-11-10.  The original PDF document was last seen online by
archive.org on 2016-03-03, and at that time wasn't different from the
first known revision.  The same revision was eventually copied to
papers/ not long ago.