BPF Exam: Implement cBPF savefile downloading.
Make the existing bytecode reading and writing code a class and extend
it to produce cBPF savefiles on demand.

Make the main HTML printing code a function; add a SELECT to the input
parameters form to provide means to download savefiles separately from
the HTML output.

Use the new class to convey the set of input parameters and the
intermediate output data between different functions; this simplifies
invocation of run_tcpdump(), run_filtertest(), r2_graph(), run_caper()
and process_request().

Refine the input parameters validation a bit more and use just one
variable to control the code flow.  Rewrite request processing to use
the new class and to produce the new result types.