Lose the "keywords" <META>.
AFAIK, in present-day Internet "keywords" is mostly a no-op: browsers do
not use it, and search engines have been using a number of advanced
methods to extract and process information from web pages, so in any
remaining cases where keywords still are an input to indexing, their
weight should be negligible.  This used to be different before spam.