Fix regen_html_pages.sh for non-GNU sed.
As it turns out, since my commit b638f15 in 2022 the script started to
generate the files incorrectly on various non-Linux hosts because "\|"
and "\+" have special meaning in GNU sed BRE, but not in other sed
implementations.

The ERE solution space is consistent in this regard across all sed
implementations that support it, but not all sed implementations support
ERE, so switching to ERE would require a test for ERE support and
possibly a way to specify the command to use GNU sed.

Keep the solution simple for the user by using "\{1,\}" instead of "+"
and multiple "s" commands instead of "|".  Update the list of tested
OSes.