regen_man_pages.sh: Fix sed backrefs for bash.
As Francois-Xavier notes, the script produces incorrect output in the
generated files when it runs using bash:

-<B><A HREF="./pcap-filter.7.html">pcap-filter</A></B>(7)
+<B><A HREF="./\2.\1.html">pcap-filter</A></B>(7)

It used to work for me because on my Linux systems /bin/sh tends to be a
symlink to /bin/dash. Due to a few implementation inconsistencies,
printing a sed regexp back-reference in a portable way using echo is not
as simple as it seems, for both the shell builtin and the external
command:

dash:
$ echo "_\\\\1_" "_\\1_" "_\1_" '_\\\\1_' '_\\1_' '_\1_'
_\1_ __ __ _\\1_ _\1_ __

bash, FreeBSD sh and NetBSD sh:
$ echo "_\\\\1_" "_\\1_" "_\1_" '_\\\\1_' '_\\1_' '_\1_'
_\\1_ _\1_ _\1_ _\\\\1_ _\\1_ _\1_

OpenBSD sh and illumos sh:
$ echo "_\\\\1_" "_\\1_" "_\1_" '_\\\\1_' '_\\1_' '_\1_'
_\1_ _\1_ _\1_ _\\1_ _\1_ _\1_

illumos sh:
$ /bin/echo "_\\\\1_" "_\\1_" "_\1_" '_\\\\1_' '_\\1_' '_\1_'
_\1_ _\1_ _\1_ _\\1_ _\1_ _\1_

all other shells:
$ /bin/echo "_\\\\1_" "_\\1_" "_\1_" '_\\\\1_' '_\\1_' '_\1_'
_\\1_ _\1_ _\1_ _\\\\1_ _\\1_ _\1_

However, single/double-quoted strings themselves and printf work
consistently everywhere, as far as I see:

$ printf '%s %s %s %s %s %s\n' \
  "_\\\\1_" "_\\1_" "_\1_" '_\\\\1_' _\\1_' '_\1_'
_\\1_ _\1_ _\1_ _\\\\1_ _\\1_ _\1_
$ printf '_\\\\1_ _\\1_ _\1_\n'
_\\1_ _\1_ __
$ printf "_\\\\1_ _\\1_ _\1_\n"
_\1_ __ __

Convert the offending echo commands to printf to fix the problem.