Improve the description of LINKTYPE_A429.
A "label" appears to be a *part* of the 32-bit words that are
LINKTYPE_A429 packets.

THere's no indication of what the byte order is for the word, but, as
the octets of the word appear to be transmitted in little-endian order
(the *bits* might be transmitted in big-endian order, but the four
*octets* appear to be transmitted with the low-order octet first), claim
that it's little-endian.  If that's not the case, feel free to report an
issue, and we can fix it.