restyle.php: Fix processing of sub-headers.
At the top of some man pages removal of "Section" and/or insertion of
"View in plain text" did not work for some man pages (tcpdump, pcap,
pcap_can_set_rfmon) or only specific versions thereof (tcpslice,
pcap_activate).  Fix this by moving "Index" out of the way of regexps
and recognizing both 0 and 3 as a valid first digit of a two-digit month
day.  Add some comments and place each resulting sub-header on a
separate line of HTML source.