Have two more sections in related.html.
This should be the last round of restructuring required for this page:
now instead of a single long "everything software" list the page has a
number of focused lists followed by a short "everything else" list.
While at it, rearrange some newlines and remove "submitted by" to make
the style consistent, get an HTML comment right and add a missing
closing tag.