restyle.php: Normalize the URLs for HTML man pages.
Apparently, some requests for the HTML man pages come with a query
string (e.g. "/manpages/page.html?ref=somesite"), some come with excess
slashes (e.g. "/manpages///page.html") and some come with unnecessary
percent-encoding.  In each case restyle.php responds with an HTTP 400.
This is a regression, compared to serving the same pages without the
script, in which case the request quirks do not prevent the server from
producing a correct response.

That said, an indefinite number of alias URLs for the same document
lowers the SNR, so instead of immediately responding with the content
redirect unusual, but otherwise valid, man page requests to their
nominal URLs.  Do not use RedirectMatch because it does not have the
query string as an input.  Do not use an additional RewriteRule because
it can handle neither "page.html?" nor repeated slashes easily.  Check
for and use REDIRECT_URL in the PHP script.  Drop the check for
REQUEST_URI because PHP always sets it.