LINKTYPE_CAN_SOCKETCAN: fix some issues.
Pick up soem fixes from
https://github.com/the-tcpdump-group/tcpdump-htdocs/pull/40:

* make clear that RTR and error message frames are only CAN CC frames
* fix typos: Classical CAN and CANXL_SEC.

Note the field that, in CAN CC frames, contains the Len 8 DLC. Describe
that value by referring to section 8.4.2.4 "DLC field" of ISO
11898-1:2015.

Put a procedure for determining the frame type, based on the one
proposed for Wireshark in
https://gitlab.com/wireshark/wireshark/-/merge_requests/18710#note_2305575009:
(which is essentially the same as the current algorithm, with some added
checks), as the second paragraph.  It's written in terms of raw octets
of the frame and raw bits as hex values, as the frame formats and flag
values are described later. It replaces an earlier description of the
first step of that procedure.

Discard some descriptions of heuristics to handle fromes from old
versions of the Linux kernel and libpcap, as those aren't used by
Wireshark, and shouldn't be necessary for dealing with the vast majority
of captures (some old captures may need fixing).

Improve the descriptions of what software producing these frames must
do.  ***NOTE:*** this does *nnot* refer to what the Linux kernel does or
should do; it refers to what programs that provide packets in the
LINKTYPE_CAN_SOCKETCAN/DLT_CAN_SOCKETCAN format must do.  For example,
this describes what libpcap must do if, for example, it's receiving
SocketCAN frames from a Linux PF_PACKET socket.

Note that remote retransmission requests may have a payload or a Len 8
DLC value that indicates a non-empty payload, but that the payload
should be ignored.