Refine reporting and file management in regen_html_pages.sh.
Instead of the "unsaved changes were lost" and "regenerated" messages,
which in most cases appear together and regardless of whether anything
has changed since the last invocation of the script, print "updated" for
files that the current invocation has changed.  In regenerate_pages()
manage a temporary file (which in many cases is on tmpfs) and in
regenerate_page() use it to avoid unnecessary file churn in the working
directory (which usually is on a regular file system).  Remove AIX,
which does not have mktemp(1), from the list of OSes.