Clean up LINKTYPE_RTAC_SERIAL.
Consistently speak of octet(s) rather than byte(s).

For DNP3, link both to the DNP home page and to the IEEE DNP3 protcol
specification.

For Modbus, link to a page closer to the specificaton.

For Synchrophasor, link to the current IEEE/IEC specification.

For SELFM, indicate what it is (even if Schweitzer Engineering
Laboratories, the "SEL" of "SELFM", doesn't appear to have put up a
specification for it).