LINKTYPE_LINUX_SLL2: Remove duplicated paragraphs
The paragraphs describing the effects of ARPHRD_IPGRE and
ARPHRD_IEEE80211_RADIOTAP on the protocol type field and payload
are duplicated. They appear once after the description of the
link-layer address and before the payload, in the middle of the
description of the effects of several other ARPHRD_ types, and
once after the description of the packet type field and before
the link-layer address length field (neither of which are related.)

There is a case, perhaps, for listing them either after the protocol
type field or ARPHRD_ type instead. LINKTYPE_LINUX_SLL makes this
somewhat easier because the protocol type field and payload are
adjacent fields, so the descriptions can occur at the end of one
field and the beginning of the next.

Signed-off-by: John Thacker <johnthacker@gmail.com>