Improve the Broadcom 4-octet DSA tag description.
For ingress and egress explain in more detail how packets travel from
external ports to the CPU and back and when exactly they are supposed to
have (or not) the DSA tag.  In the diagram captions ingress and egress
have been swapped around for a long time, get that right.  Document the
Reason code field as a bitmap to match the working code in tcpdump.
Document the Source port number filed a bit better.  Split the page into
three sections.

[skip ci]