BPF Exam: Switch to newer tcpdump & libpcap versions.
All supported Linux distributions now have tcpdump 4.99.x, so in
run_tcpdump() instead of the "-r" workaround just use "-y" and "-s"
directly.  Add libpcap 1.10.6 to the list of known versions, make it the
default and relabel the snapshot version as "1.11.0-PRE-GIT" for clarity
(this also produces a bit more useful comment in the generated .cbpf
files).